Side bar now navigates to different pages*
Side bar buttons load the same component but the component dynamically decides what data to display depending on which route is currently selected. There is a json (project-data.json) that holds the data for each project. I want to move the logic from project.component.ts to its own service so that the component is not dealing with data manipulation directly